fix(cli): move analytics arg-mapper tests to the new sibling sidecar
The dispatch_surface split (ef0a36d) left dispatch_surface_gap_tests.rs
referencing the moved *Args types via `use super::*`, which broke the bin
test target under cargo's --all-targets (caught by CI clippy/Tests, missed
locally by `clippy --quiet` which skips test compilation). Move the analytics
arg-mapper tests into dispatch_surface_analytics_tests.rs (wired into the new
module) and trim the gap test file to its remaining state/graph cases.

Verified with `cargo clippy --all-targets` (clean) and the bin test target.
